Fix some articles not rendering
Earlier, axios request was using Heading to fetch data from database
and Heading was taken from the url string passed in anchor tag which was
working fine for most articles.

But as soon as there's some special character like ' or : then the page
was not rendered because heading string didnot treat these special
characters as string.

So, now article's ID is sent through anchor tag and fetch request is
done with that ID and then the Heading from the response is used to
update the URL of the current article.